+++ src/extension.ts
@@ -136,22 +136,22 @@ export function activate(context: vscode.ExtensionContext) {
 			(tenantTreeItem) => treeManager.removeTenant(tenantTreeItem)));
 	context.subscriptions.push(
 		vscode.commands.registerCommand(commands.AGGREGATE,
-			(tenantTreeItem) => treeManager.aggregateSource(tenantTreeItem)));
+			(sourceTreeItem) => treeManager.aggregateSource(sourceTreeItem)));
 	context.subscriptions.push(
 		vscode.commands.registerCommand(commands.AGGREGATE_DISABLE_OPTIMIZATION,
-			(tenantTreeItem) => treeManager.aggregateSource(tenantTreeItem, true)));
+			(sourceTreeItem) => treeManager.aggregateSource(sourceTreeItem, true)));
 	context.subscriptions.push(
 		vscode.commands.registerCommand(commands.AGGREGATE_ENTITLEMENTS,
-			(tenantTreeItem) => treeManager.aggregateSource(tenantTreeItem, false, "entitlements")));
+			(sourceTreeItem) => treeManager.aggregateEntitlements(sourceTreeItem)));
 	context.subscriptions.push(
 		vscode.commands.registerCommand(commands.RESET_SOURCE,
-			(tenantTreeItem) => treeManager.resetSource(tenantTreeItem)));
+			(sourceTreeItem) => treeManager.resetSource(sourceTreeItem)));
 	context.subscriptions.push(
 		vscode.commands.registerCommand(commands.RESET_SOURCE_ACCOUNTS,
-			(tenantTreeItem) => treeManager.resetSource(tenantTreeItem, "entitlements")));
+			(sourceTreeItem) => treeManager.resetAccounts(sourceTreeItem)));
 	context.subscriptions.push(
 		vscode.commands.registerCommand(commands.RESET_SOURCE_ENTITLEMENTS,
-			(tenantTreeItem) => treeManager.resetSource(tenantTreeItem, "accounts")));
+			(tenantTreeItem) => treeManager.resetEntitlements(tenantTreeItem)));
 	const testConnectionCommand = new TestConnectionCommand(tenantService);
 	context.subscriptions.push(
 		vscode.commands.registerCommand(commands.TEST_SOURCE,
@@ -327,7 +327,7 @@ export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(
 		vscode.commands.registerCommand(commands.VALIDATE_CONNECTOR_RULE,
 			newConnectorRuleCommand.validateScript, newConnectorRuleCommand));
-	
+
 	const openScriptCommand = new OpenScriptCommand()
 	context.subscriptions.push(
 		vscode.commands.registerCommand(commands.EDIT_CONNECTOR_RULE,
@@ -464,15 +464,17 @@ export function activate(context: vscode.ExtensionContext) {
 			identityFilterCommand.execute, identityFilterCommand));
 	context.subscriptions.push(
 		vscode.commands.registerCommand(commands.IDENTITIES_ATT_SYNC,
-			newIdentityCommand.attSyncIdentity, newIdentityCommand));	
+			newIdentityCommand.attSyncIdentity, newIdentityCommand));
 	context.subscriptions.push(
 		vscode.commands.registerCommand(commands.IDENTITIES_DELETE,
-			newIdentityCommand.deleteIdentity, newIdentityCommand));	
+			newIdentityCommand.deleteIdentity, newIdentityCommand));
 	context.subscriptions.push(
 		vscode.commands.registerCommand(commands.IDENTITIES_PROCESS,
-			newIdentityCommand.processIdentity, newIdentityCommand));	
+			newIdentityCommand.processIdentity, newIdentityCommand));
 
 
+	// Add global interceptor for axios, to applied with the sailpoint SDK
+	// Add a request interceptor
 	axios.interceptors.request.use(onRequest)
 
 	// Add a response interceptor